move internal types into namespace so public APIs are visible at a glance
