+++ Framework/NewItems.cs
@@ -1,10 +1,10 @@
-﻿using Microsoft.Xna.Framework;
+﻿using System;
+using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using StardewModdingAPI;
 using StardewValley;
-using System;
 
-namespace Cobalt
+namespace Cobalt.Framework
 {
     internal class ItemInjector : IAssetEditor
     {
@@ -29,8 +29,8 @@ namespace Cobalt
                 asset.ReplaceWith(newTex);
                 asset.AsImage().PatchImage(oldTex);
 
-                asset.AsImage().PatchImage(Mod.instance.Helper.Content.Load<Texture2D>("cobalt-bar.png"), null, imageRect(CobaltBarItem.INDEX));
-                asset.AsImage().PatchImage(Mod.instance.Helper.Content.Load<Texture2D>("cobalt-sprinkler.png"), null, imageRect(CobaltSprinklerObject.INDEX));
+                asset.AsImage().PatchImage(ModEntry.instance.Helper.Content.Load<Texture2D>("cobalt-bar.png"), null, imageRect(CobaltBarItem.INDEX));
+                asset.AsImage().PatchImage(ModEntry.instance.Helper.Content.Load<Texture2D>("cobalt-sprinkler.png"), null, imageRect(CobaltSprinklerObject.INDEX));
             }
             else if (asset.AssetNameEquals("Data\\ObjectInformation"))
             {
